+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -440,6 +440,7 @@ sub build_isf_public_interface_contract {
             't/1271-isf-enum-member-activation-params.t',
             't/1272-isf-enum-member-rule-values.t',
             't/1273-isf-enum-member-rule-expression-values.t',
+            't/1274-isf-enum-member-rule-guard-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',